==> 01_dataset_prepare.Rmd <==
---
title: "Project - DB preparations"
author: "Andrey Kravets"
date: "`r Sys.Date()`"
output: 
   html_document:
       toc: true
       toc_float:
           collapsed: false
           smooth_scroll: true
       theme: flatly
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(rstatix)
library(pROC)
library(gtsummary)
library(readr)
library(dplyr)
```

```{r}
database = read_tsv("data/raw_database.tsv")
```
# Доделываем БД
## Отбираем колонки
```{r}
columns_of_interest <- c(
  "week_end",
  "WEEK",
  "type",
  "hosp",
  "year",
  "Направлено.всего.с.диагнозом.ОНМК",
  "Не.подтвержден.диагноз.ОНМК",
  "Госпитализировано.с.подтвержденным.диагнозом.ОНМК",
  "Пролечено.с.ОНМК.Всего....закрыто.и.б..не.считая.переводов.",
  "Внутригоспитальные.ОНМК.Всего..находились.на.лечении.в.стационаре.с.другой.патологией.",
  "с.ИИ..ишемический.инсульт.",
  "с.ТИА..транзиторно.ишемическая.атака.",
  "с.ГИ..геморрагический.инсульт..всего",
  "X..ВМК.ВЖК..внутримозговое.внутрижелудочковое.кровоизлияние..на.фоне.ГБ",
  "X..САК..субарахноидальное.кровоизлияние..при.неподтвержденной.аневризме",
  "X..АВМ..кровоизлияние.при.артерио.венозной.мальформации.",
  "X..САК.при.аневризме.мозговой.артерии",
  "Поступили.в..терапевтическое.окно.....4.5ч.",
  "из.них.ИИ",
  "из.них.ГИ",
  "Переведены.из.ПСО.в.РСЦ.для.хирургического.лечения",
  "X..ГИ.в.срок.до.48.ч..с.целью.краниотомии.",
  "X..ИИ.в.срок.до.12.ч..с.целью.ТЭ.или.стентирования.церебральных.артерий.",
  "X..ИИ.в.срок.до.2.нед..с.целью.хирургической.реваскуляризации.сосудов.шеи...КЭАЭ..КАС..ЭИКМА..АС.АП.ПА.",
  "Направлены.на.реабилитацию",
  "X..в.другой.стационар..II.этап.реабилитации.",
  "X..в.санаторий..III.этап.реабилитации.",
  "X..на.паллиативное.лечение",
  "X..на.койки..сестринского.ухода.",
  "X..на.амбулаторное.лечение..дневной.стационар.поликлиники.",
  "X..в.реабилитационное.отделение.собственного.стационара..II.этап.",
  "X..до.45.лет",
  "X..45.60.лет",
  "X..старше.60.лет",
  "X..менее.4.баллов",
  "X..более.21.балла",
  "X..от.5.до.20.баллов",
  "Тромболизис",
  "Тромбэкстракция",
  "Комбинированная.реперфузия..ТЛТ...ТЭ...до.12.часов.",
  "Реваскуляризирующие.операции.в.остром.периоде..всего",
  "При.ГИ..всего",
  "Летальные..исходы..ЛИ..всего",
  "всего.ЛИ.в.1.сутки.поступления",
  "Выписано.из.стационара.всего",
  "из.них.по.шкале.Ривермид....1.балл",
  "из.них.по.шкале.Ривермид..от.2.до.4.баллов",
  "из.них.по.шкале.Ривермид..от.5.до.8.баллов",
  "из.них.по.шкале.Ривермид...9.баллов"
)

database_selected <- database[, columns_of_interest]

rename_columns <- c(
  "Направлено.всего.с.диагнозом.ОНМК" = "Всего_с_ОНМК",
  "Не.подтвержден.диагноз.ОНМК" = "Не_подтвержден_ОНМК",
  "Госпитализировано.с.подтвержденным.диагнозом.ОНМК" = "Госпитализированы_с_ОНМК",
  "Пролечено.с.ОНМК.Всего....закрыто.и.б..не.считая.переводов." = "Пролечены_с_ОНМК_Всего",
  "Внутригоспитальные.ОНМК.Всего..находились.на.лечении.в.стационаре.с.другой.патологией." = "Внутрибольничные_ОНМК",
  "с.ИИ..ишемический.инсульт." = "ИИ_Ишемический_Инсульт",
  "с.ТИА..транзиторно.ишемическая.атака." = "ТИА_Транзиторная_Ишемическая_Атака",
  "с.ГИ..геморрагический.инсульт..всего" = "ГИ_Геморрагический_Инсульт",
  "X..ВМК.ВЖК..внутримозговое.внутрижелудочковое.кровоизлияние..на.фоне.ГБ" = "ВМК_ВЖК",
  "X..САК..субарахноидальное.кровоизлияние..при.неподтвержденной.аневризме" = "САК_Неподтвержденная_Аневризма",
  "X..АВМ..кровоизлияние.при.артерио.венозной.мальформации." = "Кровоизлияние_АВМ",
  "X..САК.при.аневризме.мозговой.артерии" = "САК_Аневризма",
  "Поступили.в..терапевтическое.окно.....4.5ч." = "Терапевтическое_Окно_4_5ч",
  "из.них.ИИ" = "ИИ_Терапевтическое_Окно",
  "из.них.ГИ" = "ГИ_Терапевтическое_Окно",
  "Переведены.из.ПСО.в.РСЦ.для.хирургического.лечения" = "Переведены_ПСО_РСЦ",
  "X..ГИ.в.срок.до.48.ч..с.целью.краниотомии." = "ГИ_Переведены_до_48ч",
  "X..ИИ.в.срок.до.12.ч..с.целью.ТЭ.или.стентирования.церебральных.артерий." = "ИИ_Переведены_до_12ч",
  "X..ИИ.в.срок.до.2.нед..с.целью.хирургической.реваскуляризации.сосудов.шеи...КЭАЭ..КАС..ЭИКМА..АС.АП.ПА." = "ИИ_Переведены_до_2_недель",
  "Направлены.на.реабилитацию" = "Направлены_на_Реабилитацию",
  "X..в.другой.стационар..II.этап.реабилитации." = "II_Этап_Реабилитация",
  "X..в.санаторий..III.этап.реабилитации." = "III_Этап_Реабилитация",
  "X..на.паллиативное.лечение" = "Паллиативное_Лечение",
  "X..на.койки..сестринского.ухода." = "Сестринский_Уход",
  "X..на.амбулаторное.лечение..дневной.стационар.поликлиники." = "Амбулаторное_Лечение",
  "X..в.реабилитационное.отделение.собственного.стационара..II.этап." = "Собственная_Реабилитация_II_Этап",
  "X..до.45.лет" = "Младше_45_лет",
  "X..45.60.лет" = "От_45_до_60_лет",
  "X..старше.60.лет" = "Старше_60_лет",
  "X..менее.4.баллов" = "NIHSS_Меньше_4",
  "X..более.21.балла" = "NIHSS_Больше_21",
  "X..от.5.до.20.баллов" = "NIHSS_5_20",
  "Тромболизис" = "Тромболизис",
  "Тромбэкстракция" = "Тромбэкстракция",
  "Комбинированная.реперфузия..ТЛТ...ТЭ...до.12.часов." = "Комбинированная_Реперфузия",
  "Реваскуляризирующие.операции.в.остром.периоде..всего" = "Реваскуляризация_Острый_Период",
  "При.ГИ..всего" = "ГИ_Всего",
  "Летальные..исходы..ЛИ..всего" = "Летальные_Исходы",
  "всего.ЛИ.в.1.сутки.поступления" = "Летальные_Исходы_1_Сутки",
  "Выписано.из.стационара.всего" = "Выписаны_Всего",
  "из.них.по.шкале.Ривермид....1.балл" = "Ривермид_1_Балл",
  "из.них.по.шкале.Ривермид..от.2.до.4.баллов" = "Ривермид_2_4_Балла",
  "из.них.по.шкале.Ривермид..от.5.до.8.баллов" = "Ривермид_5_8_Баллов",
  "из.них.по.шкале.Ривермид...9.баллов" = "Ривермид_9_Баллов"
)

database_selected <- database_selected |> 
  rename_with(~ ifelse(!is.na(rename_columns[.x]), rename_columns[.x], .x))
```

## Даты
```{r}
#if week_end has 3 characters, add 0 to the beginning - to format it as ddmmYYYY


database_selected <- database_selected |>
  mutate(week_end = as.character(week_end)) %>%
  mutate(week_end = if_else(str_length(as.character(week_end)) == 3, paste0("0", week_end), week_end),
         date = paste0(week_end, year), 
         date = as.Date(date, format = "%d%m%Y")) |>
  select(date, year, everything()) |>
  select(-week_end, -WEEK)

database_selected_final <- database_selected |>
  group_by(hosp, type, date) |>
  mutate(week_num = row_number()) |>
  ungroup()

database_selected_final <- database_selected_final |>
  select(week_num, date, everything())
head(database_selected_final)
```

## Сначала засуммируем др и невр
Видим, что в др нет "Всего_с_ОНМК" => значит можем засуммировать невр и др (type 1 и type 2)  

```{r}
database_selected_final %>% filter(type == 2) %>% select(starts_with('Всего_с_ОНМК')) %>%
  filter(!is.na(`Всего_с_ОНМК`))

database_selected_final
```

```{r}
#part = database_selected %>% filter(hosp %in% c("b3", "alm")) 

database_summarized = database_selected_final %>% group_by(hosp, year, date) %>%
  summarise(across(everything(), sum, na.rm = TRUE), .groups = "drop")

database_summarized
```
```{r}
database_summarized %>% filter(type != 3)
```

## Добавляем по месяцам
```{r}
cols_to_sum = colnames(database_summarized)[6:49]

database_summarized_month <- database_summarized %>%
  mutate(
    # Создаём столбец month, группируя недели по 4, но не превышая 12
    month = month(date),
    # Преобразуем первые 5 столбцов в факторы, чтобы они не участвовали в суммировании
    across(1:5, ~ as.factor(.x))) %>%
  group_by(hosp, type, year, month) %>%
  summarize(across(
      all_of(cols_to_sum),
      ~ if (all(is.na(.x))) NA_real_ else sum(.x, na.rm = TRUE)),
    .groups = "drop") %>%
  mutate(month = as.factor(month))

database_summarized_month
```

```{r}
database_summarized_month <- database_summarized_month %>%
  mutate(across(where(is.numeric), ~ ifelse(rowSums(across(where(is.numeric)) == 0) == ncol(across(where(is.numeric))), NA, .)))

# Print the updated table
database_summarized_month
```
## Суммирую загрузку
```{r}
database_summarized_month = database_summarized_month %>%
  mutate(`Суммарная_Загрузка_ОНМК` = `Всего_с_ОНМК`,
         `Суммарно_Госпитализировано` = `Госпитализированы_с_ОНМК`)
```

## Fancy Graphs
```{r}
library(visdat)
vis_dat(database_summarized_month)
```
```{r}
missing_data <- database_summarized_month %>%
  mutate(is_missing = ifelse(is.na(`Суммарная_Загрузка_ОНМК`), 1, 0)) %>%
  group_by(hosp, month) %>%
  summarise(missing_count = sum(is_missing, na.rm = TRUE), .groups = "drop")

# Plot the missing data by `hosp` and `month`
ggplot(missing_data, aes(x = month, y = hosp, fill = missing_count)) +
  geom_tile(color = "white") +
  scale_fill_gradient(low = "white", high = "red", name = "Missing Count") +
  labs(
    title = "Missing Data by Hosp and Month by All Years",
    x = "Month",
    y = "Hosp"
  ) +
  theme_minimal()
```

```{r}
write_tsv(database_summarized_month, "data/data_summarized_by_month.tsv")
```
