==> 04_trend_analysis_full.Rmd <==
---
title: "Trend analysis"
author: "Victoria Zaitceva"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
library(tidyverse)
library(ggpubr)
library(tidyr)
library(anytime)
library(lubridate)
library(zoo)
library(patchwork)

data <- read.table("data/data_summarized_by_month_filled_hosp_MR.tsv", header = TRUE, sep = "\t")

theme_custom <- theme(
    panel.background = element_rect(fill = "white"),
    plot.title = element_text(size = 15, hjust = 0.5),
    plot.subtitle = element_text(size = 12, hjust = 0.5),
    strip.text = element_text(size = 10),
    axis.text = element_text(size = 10),
    axis.text.x = element_text(angle = 90, hjust = 1), #rotate
    axis.title = element_text(size = 10),
    legend.title = element_text(size = 9),
    legend.text = element_text(size = 9)
  )

Sys.setlocale("LC_ALL", "en_US.UTF-8")  # to avoid sum(\u41f" error

```


Map types of the hospitals (it is all 3 for some reason)
We might use the type for procedures


```{r}
hosp_to_type <- c(
  b17 = 2, b3 = 2, b16 = 2, bsp = 2, b26 = 2, mcs = 2, b2 = 2, alm = 2,
  b1 = 1, b40 = 2, gvv = 2, b38 = 1, b15 = 1, b4 = 1, b33 = 1, b37 = 1
)


data$type <- hosp_to_type[data$hosp]
```



# Годовая динамика в структуре заболевания (%) общегородская

Первый график - просто показать, почему я не хотела объединять все в один график.


```{r}

stroke_types <- c("ИИ_Ишемический_Инсульт", "ГИ_Геморрагический_Инсульт", "ТИА_Транзиторная_Ишемическая_Атака")
# Combine Data for All Stroke Types
data_combined <- data %>%
  pivot_longer(
    cols = stroke_types,
    names_to = "Stroke_Type",
    values_to = "Count"
  ) %>%
  group_by(year, Stroke_Type) %>%
  summarise(
    Total_Stroke = sum(Count, na.rm = TRUE),
    Total_Cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
    .groups = "drop"
  ) %>%
  mutate(
    Percentage = Total_Stroke / Total_Cases * 100
  )

# Plot All Stroke Types on One Plot
final_plot <- ggplot(data_combined, aes(x = year, y = Percentage, color = Stroke_Type)) +
  geom_point() +
  geom_line(size = 1) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +
  labs(
    title = "Динамика в структуре заболевания (%) по годам общегородская",
    x = "Год",
    y = "% от общего числа пролеченных"
  ) +
  theme_custom +
  theme(
    legend.title = element_text(size = 12, face = "bold"),
    legend.position = "top",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 10)
  )

# Print the Plot
print(final_plot)

```




```{r, fig.width = 12, fig.height = 2.5}

stroke_types <- c("ИИ_Ишемический_Инсульт", "ГИ_Геморрагический_Инсульт", "ТИА_Транзиторная_Ишемическая_Атака")

plots <- list()



for (type in stroke_types) {
  

  temp <- data %>%
    group_by(year) %>%
    summarise(
      !!sym(type) := sum(!!sym(type), na.rm = TRUE),
      Пролечены_с_ОНМК_Всего = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
      .groups = "drop"
    ) %>%
    mutate(
      Percentage = !!sym(type) / Пролечены_с_ОНМК_Всего * 100
    ) %>%
    group_by(year) %>%
    summarise(
      Percentage = mean(Percentage, na.rm = TRUE),
      .groups = "drop"
    )
  
  

  plot <- ggplot(temp, aes(x = year, y = Percentage)) +
    geom_rect(
      aes(xmin = 2020, xmax = 2022, ymin = -Inf, ymax = Inf), 
      fill = "red", alpha = 0.01  # Light red transparent highlight for COVID period
    ) +
    geom_point() +
    geom_line() +
    scale_x_continuous(breaks = seq(2017, 2024, by = 1))+
    scale_y_continuous(labels = scales::percent_format(scale = 1)) +
    labs(
      title = paste(type),
      x = NULL,
      y = NULL
    ) +
    theme_custom +
    theme(
      plot.title = element_text(size = 12))
  

  plots[[type]] <- plot
}



combined_plot <- ggarrange(
  plotlist = plots, 
  ncol = 3, 
  nrow = 1,
  widths = c(1,1,1), 
  common.legend = TRUE,  
  labels = NULL          
)



final_plot <- annotate_figure(
  combined_plot,
  top = text_grob("Disease structure trend", size = 14, face = "bold"), 
  left = text_grob("% of total treated", size = 10, face = "plain", rot = 90)   
)


print(final_plot)

ggsave("images/structure_trend_year.png", final_plot, width = 12, height = 4)
```

# Месячная динамика в структуре заболевания (%) общегородская

```{r, fig.width = 12, fig.height = 4}

plots <- list()



for (type in stroke_types) {
  

  temp <- data %>%
    mutate(year_month = anytime::anydate(paste0(year, paste0(ifelse(month < 10, paste0("0", month), month), "01"), sep = "-"))) %>%
    group_by(year_month) %>%
    summarise(
      !!sym(type) := sum(!!sym(type), na.rm = TRUE),
      Пролечены_с_ОНМК_Всего = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
      .groups = "drop"
    ) %>%
    mutate(
      Percentage = !!sym(type) / Пролечены_с_ОНМК_Всего * 100
    ) %>%
    group_by(year_month) %>%
    summarise(
      Percentage = mean(Percentage, na.rm = TRUE),
      .groups = "drop"
    )
  
  temp$year_month_num <- as.numeric(as.Date(temp$year_month))
  

  plot <- ggplot(temp, aes(x = year_month, y = Percentage)) +
    annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1  # Adjust transparency level
    ) +
    geom_point(size = 1) +
    geom_line() +
    geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
    labs(
      title = paste(type),
      x = NULL,
      y = NULL
    ) +
    scale_x_date(
    date_breaks = "4 month",      
    date_labels = "%Y-%m"          
  ) +
    scale_y_continuous(labels = scales::percent_format(scale = 1)) +
    theme_custom +
    theme(
      plot.title = element_text(size = 12),
      axis.text.x = element_text(hjust = 1, size = 8))
  

  plots[[type]] <- plot
}


combined_plot <- ggarrange(
  plotlist = plots, 
  ncol = 3, 
  nrow = 1,
  widths = c(1,1,1), 
  common.legend = TRUE,  
  labels = NULL          
)



final_plot <- annotate_figure(
  combined_plot,
  top = text_grob("Disease structure dynamics by month", size = 14, face = "bold"),  # Common title
  bottom = text_grob("", size = 12, face = "plain"),  # Common x-axis title
  left = text_grob("% of total treated", size = 10, face = "plain", rot = 90)  # Common y-axis title
)


print(final_plot)

ggsave("images/structure_trend_month.png", final_plot, width = 12, height = 4)

```






# Тромболизис


Тромболизис / ИИ_Ишемический_Инсульт

## 

```{r}
data_prepared <- data %>%
  group_by(year) %>%
  summarise(
    total_thromb = sum(Тромболизис, na.rm = TRUE), 
    total_cases = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_thromb / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year, y = Летальные_Исходы_проц)) +
  geom_line(color = "black", size = 0.7) +  
  geom_point(color = "black", size = 0.8) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1))+ 
  labs(
    title = "Thrombolysis",
    x = "",
    y = "% from total IS cases"
  ) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 10)
  )
```

## Месячная динамика в % выполненного тромболизиса


```{r}
data_prepared <- data %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(year_month) %>%
  summarise(
    total_thromb = sum(Тромболизис, na.rm = TRUE), 
    total_cases = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_thromb / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1  
    ) +
  geom_line(color = "black", size = 0.6) +  
  geom_point(color = "black", size = 1.2) +
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkgreen") +
  scale_x_date(date_breaks = "3 months", 
               date_labels = "%Y-%m") + 
  scale_y_continuous(labels = scales::percent_format(scale = 1)) +
  labs(
    title = "Thrombolysis rate by month",
    x = "",
    y = "% from total IS cases"
  ) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 10),
    axis.text.y = element_text(size = 12),
    # bold title
    plot.title = element_text(size = 14, face = "bold")
  )

ggsave("images/thromb_month.png", width = 6, height = 4)

```

## Годовой тренд по тромболизису в разных типах центров

```{r, fig.width = 12, fig.height = 4}
data_prepared <- data %>%
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(year_month, type) %>%
  summarise(
    total_thromb = sum(Тромболизис, na.rm = TRUE), 
    total_cases = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_thromb / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц)) +
  geom_line(color = "black", size = 0.7) +  
  geom_point(color = "black", size = 0.8) +
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1))+ 
  labs(
    title = "Thrombolysis by hosp type",
    x = "",
    y = "% from total IS cases"
  ) +
  scale_x_date(date_breaks = "3 months", 
               date_labels = "%Y-%m") + 
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 10)
  ) +
  facet_wrap(vars(type), scales = "free")  
```






## Месячные тренды по тромболизису в центрах типа 1

Тромболизис/ИИ_Ишемический_Инсульт

```{r}
data_prepared <- data %>%
  filter(type == 1) %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(hosp, year_month) %>%
  summarise(
    total_thromb = sum(Тромболизис, na.rm = TRUE), 
    total_cases = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_thromb / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  geom_line(color = "black", size = 0.5) +  
  geom_point(color = "black", size = 0.5) +
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
  scale_x_date(date_breaks = "6 months", 
               date_labels = "%Y-%m") + 
  labs(
    title = "Thrombolysis in primary vascular care centers (type 1)",
    x = "",
    y = "% from total IS cases"
  ) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 6)
  ) +
  facet_wrap(vars(hosp), scales = "free")
```
## Месячные тренды по тромболизису в центрах типа 2

```{r}
data_prepared <- data %>%
  filter(type == 2) %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(hosp, year_month) %>%
  summarise(
    total_thromb = sum(Тромболизис, na.rm = TRUE), 
    total_cases = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_thromb / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  geom_line(color = "black", size = 0.5) +  
  geom_point(color = "black", size = 0.5) +
  geom_smooth(method = "loess", span = 0.3, se = TRUE, color = "darkred") +
  scale_x_date(date_breaks = "6 months", 
               date_labels = "%Y-%m") + 
  labs(
    title = "Thrombolysis in regional vascular care centers (type 2)",
    x = "",
    y = "% from total IS cases"
  ) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 6)
  ) +
  facet_wrap(vars(hosp), scales = "free")
```
# Летальные исходы

Летальные_Исходы / Пролечены_с_ОНМК_Всего

## Годовая динамика летальных исходов

```{r}

data_aggregated <- data %>%
  group_by(year) %>%
  summarise(
    total_deaths = sum(Летальные_Исходы, na.rm = TRUE), 
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_deaths / total_cases * 100
  ) %>%
  pivot_longer(
    cols = c(Летальные_Исходы_проц),
    names_to = "type",
    values_to = "percentage"
  )


ggplot(data_aggregated, aes(x = year, y = percentage, color = type)) +
  geom_line(color = "black") + 
  geom_point(color = "black") + 
  labs(
    title = "Общегородская доля летальных исходов",
    x = "",
    y = "% from total treated"
  ) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +  
  theme_custom +
  theme(
    legend.position = "none"
  
  )

```

## Месячная динамика летальных исходов

```{r}
data_prepared <- data %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(year_month) %>%
  summarise(
    total_deaths = sum(Летальные_Исходы, na.rm = TRUE), 
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_deaths / total_cases * 100
  ) 

# Plot
plot <- ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1  
    ) +
  geom_line(color = "black", size = 0.7) +  
  geom_point(color = "black", size = 1.2) +
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
  scale_x_date(date_breaks = "3 months", date_labels = "%Y-%m") + 
  scale_y_continuous(labels = scales::percent_format(scale = 1)) +
  labs(
    title = "Mortality rate by month",
    x = "",
    y = "% from total treated"
  ) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 10),
    axis.text.y = element_text(size = 12),
    # bold title
    plot.title = element_text(size = 14, face = "bold")
  )

print(plot)

ggsave("images/mortality_month.png", width = 8, height = 4)
```

## Годовой тренд по летальным исходам в разных типах центров


```{r}
data_prepared <- data %>%
  # Create year_month column
  
  group_by(type, year) %>%
  summarise(
    total_deaths = sum(Летальные_Исходы, na.rm = TRUE), 
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_deaths / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year, y = Летальные_Исходы_проц, color = type)) +
  geom_line(color = "black") + 
  geom_point(color = "black") + 
  labs(
    title = "Общегородская доля летальных исходов",
    x = "",
    y = "%"
  ) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +  
  scale_y_continuous(expand = c(0, 0)) + # Ensure the y-axis starts at 0
  theme_custom +
  theme(
    legend.position = "none"
  ) +
  facet_wrap(vars(type), scales = "fixed")


```

## Месячные тренды по летальным исходам в центрах типа 1

b37, b4

```{r}
data_prepared <- data %>%
  filter(type == 1) %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(hosp, year_month) %>%
  summarise(
    total_deaths = sum(Летальные_Исходы, na.rm = TRUE), 
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_deaths / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1  
    ) +
  geom_line(color = "black", size = 0.4) +  
  geom_point(color = "black", size = 0.6) +
  geom_smooth(method = "loess", span = 0.3, se = TRUE, color = "darkred") +
  scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") + 
  labs(
    title = "Monthly mortality rate in hosp type 1",
    x = "",
    y = "% from total treated"
  ) +
  facet_wrap(vars(hosp), scales = "free") +  # Facet by hospital
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 4)
  )


```
## Месячные тренды по летальным исходам в центрах типа 2


b17

```{r}
data_prepared <- data %>%
  filter(type == 2) %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(hosp, year_month) %>%
  summarise(
    total_deaths = sum(Летальные_Исходы, na.rm = TRUE), 
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_deaths / total_cases * 100
  ) 

# Plot
ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1  
    ) +
  geom_line(color = "black", size = 0.4) +  
  geom_point(color = "black", size = 0.6) +
  geom_smooth(method = "loess", span = 0.3, se = TRUE, color = "darkred") +
  scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") + 
  labs(
    title = "Monthly mortality rate in hosp type 2",
    x = "",
    y = "% from total treated"
  ) +
  facet_wrap(vars(hosp), scales = "free") +  # Facet by hospital
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 4)
  )


```
Mortality rate for each hospital


```{r}
data_prepared <- data %>%
  # Create year_month column
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(hosp, year_month) %>%
  summarise(
    total_deaths = sum(Летальные_Исходы, na.rm = TRUE), 
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE), 
    .groups = "drop"
  ) %>%
  mutate(
    Летальные_Исходы_проц = total_deaths / total_cases * 100
  ) 

# Plot
plot <- ggplot(data_prepared, aes(x = year_month, y = Летальные_Исходы_проц, colour = hosp)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.2  
    ) +
  geom_line(color = "black", size = 0.4) +  
  geom_point(color = "black", size = 0.6) +
  geom_smooth(method = "loess", span = 0.3, se = TRUE, color = "darkred") +
  scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") + 
  labs(
    title = "",
    x = "",
    y = "% from total treated"
  ) +
  facet_wrap(vars(hosp), scales = "free") +  # Facet by hospital
  theme_custom +
  theme(
    legend.position = "none",
    #axis.text.x = element_text(angle = 90, hjust = 1, size = 4),
    axis.text.y = element_text(size = 6),
    #remove x-axis text
    axis.text.x = element_blank()
  )

print(plot)

ggsave("images/mortality_hosp_month.png", width = 8, height = 4)
```


# Терапевтическое окно


ИИ_Терапевтическое_Окно / ИИ_Ишемический_Инсульт
ГИ_Терапевтическое_Окно / ГИ_Геморрагический_Инсульт

## Годовые общегородские доли пациентов, попавших в ТО (4.5 часа)



```{r}

data_aggregated <- data %>%
  group_by(year) %>%
  summarise(
    total_ii_window = sum(ИИ_Терапевтическое_Окно, na.rm = TRUE),
    total_gi_window = sum(ГИ_Терапевтическое_Окно, na.rm = TRUE),
    total_ii_cases = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE),
    total_gi_cases = sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE),
    .groups = "drop"
  ) %>%
  mutate(
    ii_проц = total_ii_window / total_ii_cases * 100,
    gi_проц = total_gi_window / total_gi_cases * 100
  ) %>%
  pivot_longer(
    cols = c(ii_проц, gi_проц),
    names_to = "type",
    values_to = "percentage"
  )


ggplot(data_aggregated, aes(x = year, y = percentage, color = type)) +
  geom_line() + 
  geom_point() + 
  scale_color_manual(
    values = c("ii_проц" = "#1f77b4", "gi_проц" = "darkred"),
    labels = c("ii_проц" = "ИИ", "gi_проц" = "ГИ"),
    name = "Тип инсульта"
  ) +
  labs(
    title = "Общегородские доли пациентов, попавших в ТО (4.5 часа)",
    x = "",
    y = "%"
  ) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +  
  theme_custom +
  theme(
    legend.position = "bottom",
  )

```
## Месячные общегородские доли пациентов, попавших в ТО (4.5 часа)

```{r}

data_aggregated <- data %>%
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(year_month) %>%
  summarise(
    total_window = sum(Терапевтическое_Окно_4_5ч, na.rm = TRUE),  
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
    .groups = "drop"
  ) %>%
  mutate(
    проц = total_window / total_cases * 100
  ) %>%
  pivot_longer(
    cols = c(проц),
    names_to = "type",
    values_to = "percentage"
  )


plot <- ggplot(data_aggregated, aes(x = year_month, y = percentage, color = type)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1 
    ) +
  
  geom_line(color = "black") + 
  geom_point(color = "black") + 
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkblue") +

  labs(
    title = "Therapeutic time window dynamics by month",
    x = "",
    y = "% of total treated"
  ) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +  
  scale_x_date(date_breaks = "3 months", date_labels = "%Y-%m") + 
  scale_y_continuous(labels = scales::percent_format(scale = 1)) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 10),
    axis.text.y = element_text(size = 12),
    # bold title
    plot.title = element_text(size = 14, face = "bold")
  )

print(plot)

ggsave("images/therapeutic_window_month.png", plot, width = 6, height = 4)


```

## Месячные общегородские доли пациентов, попавших в ТО (4.5 часа) в разных типах центров


```{r}

data_aggregated <- data %>%
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(type, year_month) %>%
  summarise(
    total_window = sum(Терапевтическое_Окно_4_5ч, na.rm = TRUE),  
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
    .groups = "drop"
  ) %>%
  mutate(
    проц = total_window / total_cases * 100
  ) %>%
  pivot_longer(
    cols = c(проц),
    names_to = "metric", # don't use type - becaus etype is a variable
    values_to = "percentage"
  )


ggplot(data_aggregated, aes(x = year_month, y = percentage, color = type)) +
  geom_line(color = "black") + 
  geom_point(color = "black") + 
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +

  labs(
    title = "Общегородские доли пациентов, попавших в ТО (4.5 часа)",
    x = "",
    y = "%"
  ) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +  
  scale_x_date(date_breaks = "3 months", date_labels = "%Y-%m") + 
  theme_custom +

  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 6)
  ) +
  facet_wrap(vars(type), scales = "free")
```

## ТО в "чистых" центрах


```{r}

data_aggregated <- data %>%
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01")),
         clean = case_when(
           hosp %in% c("b3", "b26", "bsp") ~ "vascular (b3, b26, bsp)",
           TRUE ~ "reclassified"
         )) %>%
  group_by(clean, year_month) %>%  # Preserve `clean` here
  summarise(
    total_window = sum(Терапевтическое_Окно_4_5ч, na.rm = TRUE),  
    total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
    .groups = "drop"
  ) %>%
  mutate(
    проц = total_window / total_cases * 100
  ) %>%
  pivot_longer(
    cols = c(проц),
    names_to = "metric",
    values_to = "percentage"
  )

ggplot(data_aggregated, aes(x = year_month, y = percentage, color = type)) +
  annotate(
      "rect",
      xmin = as.Date("2020-03-01"), 
      xmax = as.Date("2021-12-31"), 
      ymin = -Inf, ymax = Inf, 
      fill = "red", alpha = 0.1 
    ) +
  geom_line(color = "black") + 
  geom_point(color = "black") + 
  geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkblue") +

  labs(
    title = "Therapeutic time window dynamics by COVID profile",
    x = "",
    y = "% of total treated"
  ) +
  scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +  
  scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") + 
  scale_y_continuous(labels = scales::percent_format(scale = 1)) +
  theme_custom +
  theme(
    legend.position = "none",
    axis.text.x = element_text(angle = 90, hjust = 1, size = 9),
    axis.text.y = element_text(size = 10),
    # bold title
    plot.title = element_text(size = 14, face = "bold")
  ) +
  facet_wrap(vars(clean), scales = "free")
```




# Корреляционный анализ 


## Correlation matrix


```{r}

library(ggplot2)
library(reshape2)
library(corrplot)


data_prepared <- data %>%
  mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
  group_by(year_month) %>%
  summarise(
    mortality = sum(Летальные_Исходы, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    IS_perc = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    GS_perc = sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    age60_perc = sum(Старше_60_лет, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    severety_perc = sum(NIHSS_Больше_21, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    ii_TW_perc = sum(ИИ_Терапевтическое_Окно, na.rm = TRUE) / sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) * 100,
    gi_TW_perc = sum(ГИ_Терапевтическое_Окно, na.rm = TRUE) / sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE) * 100,
    total_window = sum(Терапе, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    thromb_perc = sum(Тромболизис, na.rm = TRUE) / sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) * 100,
    total_treated = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
    total_load = sum(Суммарная_Загрузка_ОНМК, na.rm = TRUE),
    .groups = "drop"
  )
    


time_series_data <- data_prepared %>%
  select(-year_month)  # Only numeric time series columns


cor_matrix <- cor(time_series_data, use = "pairwise.complete.obs", method = "pearson")


cor_melted <- melt(cor_matrix)  
cor_melted_plot <- ggplot(cor_melted, aes(Var1, Var2, fill = value)) +
  geom_tile(color = "white") +
  geom_text(aes(label = round(value, 2)), color = "black", size = 4) +  # Add values
  scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
                       midpoint = 0, limit = c(-1, 1), space = "Lab", name = "Pearson r") +
  labs(title = "Correlation matrix (all hospitals)",
       x = "", y = "") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))


#ggsave("images/correlation_matrix.png", cor_melted_plot)

cor_melted_plot

```
## Correlation matrix for each hosp

```{r}
library(ggplot2)
library(reshape2)
library(dplyr)
library(anytime)


data_prepared <- data %>%
  mutate(
    year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))
  ) %>%
  group_by(hosp, year_month) %>%
  summarise(
    mortality = sum(Летальные_Исходы, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    IS_perc = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    GS_perc = sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    age60_perc = sum(Старше_60_лет, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    severety_perc = sum(NIHSS_Больше_21, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    ii_TW_perc = sum(ИИ_Терапевтическое_Окно, na.rm = TRUE) / sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) * 100,
    gi_TW_perc = sum(ГИ_Терапевтическое_Окно, na.rm = TRUE) / sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE) * 100,
    total_window = sum(ИИ_Терапевтическое_Окно, ГИ_Терапевтическое_Окно, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
    thromb_perc = sum(Тромболизис, na.rm = TRUE) / sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) * 100,
    total_treated = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
    total_load = sum(Суммарная_Загрузка_ОНМК, na.rm = TRUE),
    .groups = "drop"
  )

# Function to calculate and plot the upper triangular correlation matrix
plot_upper_triangle <- function(hosp_name, data) {
  # Filter data for the hospital
  hospital_data <- data %>% filter(hosp == hosp_name) %>% select(-hosp, -year_month)
  
  # Calculate correlation matrix
  cor_matrix <- cor(hospital_data, use = "pairwise.complete.obs", method = "pearson")
  
  # Mask the lower triangle
  cor_matrix[lower.tri(cor_matrix, diag = FALSE)] <- NA
  
  # Melt the correlation matrix
  cor_melted <- melt(cor_matrix, na.rm = TRUE)  # Remove NA entries
  
  # Plot
  ggplot(cor_melted, aes(Var1, Var2, fill = value)) +
    geom_tile(color = "white") +
    geom_text(aes(label = round(value, 2)), color = "black", size = 4) +  # Add values
    scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
                         midpoint = 0, limit = c(-1, 1), space = "Lab", name = "Pearson r") +
    labs(
      title = paste("Correlation matrix for", hosp_name),
      x = "", y = ""
    ) +
    theme_custom +
    theme(axis.text.x = element_text(angle = 45, hjust = 1))
}

# Create correlation matrix plots for each hospital
hospitals <- unique(data_prepared$hosp)
plots <- lapply(hospitals, function(hosp_name) {
  plot_upper_triangle(hosp_name, data_prepared)
})


plots[[1]]

#ggplot2::ggsave("images/correlation_matrix_hosp2.png", plots[[2]])

```


Correlation with significance

```{r}
library(ggplot2)
library(reshape2)
library(dplyr)

# Function to calculate correlation matrix and p-values
calc_cor_pval <- function(data) {
  var_names <- colnames(data)
  cor_matrix <- cor(data, use = "pairwise.complete.obs", method = "pearson")
  p_matrix <- matrix(NA, nrow = ncol(data), ncol = ncol(data))
  colnames(p_matrix) <- var_names
  rownames(p_matrix) <- var_names
  
  for (i in seq_len(ncol(data))) {
    for (j in seq_len(ncol(data))) {
      if (i <= j) {
        test <- cor.test(data[[i]], data[[j]], method = "pearson")
        p_matrix[i, j] <- test$p.value
        p_matrix[j, i] <- test$p.value
      }
    }
  }
  
  list(cor_matrix = cor_matrix, p_matrix = p_matrix)
}

# Calculate correlation matrix and p-values
results <- calc_cor_pval(time_series_data)
cor_matrix <- results$cor_matrix
p_matrix <- results$p_matrix

# Melt the correlation matrix and p-value matrix
cor_melted <- melt(cor_matrix)
p_melted <- melt(p_matrix)

# Combine correlations and p-values into a single data frame
cor_melted <- cor_melted %>%
  rename(correlation = value) %>%
  mutate(p_value = p_melted$value)

# Add significance labels (e.g., "*")
cor_melted <- cor_melted %>%
  mutate(
    sig_label = case_when(
      p_value < 0.01 ~ "**",
      p_value < 0.05 ~ "*",
      TRUE ~ ""
    )
  )

# Plot with significance labels
cor_melted_plot <- ggplot(cor_melted, aes(Var1, Var2, fill = correlation)) +
  geom_tile(color = "white") +
  geom_text(aes(label = paste0(round(correlation, 2), sig_label)), color = "black", size = 5) +  # Add values with significance
  scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
                       midpoint = 0, limit = c(-1, 1), space = "Lab", name = "Pearson r") +
  labs(title = "",
       x = "", y = "") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 15),
        axis.text.y = element_text(size = 15))

cor_melted_plot

ggsave("images/correlation_matrix_significance.png", cor_melted_plot, width = 10, height = 8)

```
