==> 05_Shiny.Rmd <==
---
title: "Shiny App"
author: "Victoria Zaitceva"
date: "`r Sys.Date()`"
output: html_document
runtime: shiny
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
```

```{r}

library(shiny)
library(tidyverse)
library(ggpubr)
library(anytime)
library(lubridate)
library(patchwork)
library(reshape2)
library(corrplot)



load_data <- function(filepath) {
  read.table(filepath, header = TRUE, sep = "\t")
}


data <- load_data("data/data_summarized_by_month_filled_hosp_MR.tsv")



theme_custom <- theme(
  panel.background = element_rect(fill = "white"),
  plot.title = element_text(size = 15, hjust = 0.5),
  plot.subtitle = element_text(size = 12, hjust = 0.5),
  strip.text = element_text(size = 10),
  axis.text.y = element_text(size = 12),
  axis.text.x = element_text(size = 11, angle = 90, hjust = 1),
  axis.title = element_text(size = 12),
  legend.title = element_text(size = 9),
  legend.text = element_text(size = 9)
)

# hospital types
hosp_to_type <- c(
  b17 = 2, b3 = 2, b16 = 2, bsp = 2, b26 = 2, mcs = 2, b2 = 2, alm = 2,
  b1 = 1, b40 = 2, gvv = 2, b38 = 1, b15 = 1, b4 = 1, b33 = 1, b37 = 1
)

data$type <- hosp_to_type[data$hosp]


# calculate correlation matrix and p-values
calc_cor_pval <- function(data) {
  var_names <- colnames(data)
  cor_matrix <- cor(data, use = "pairwise.complete.obs", method = "pearson")
  p_matrix <- matrix(NA, nrow = ncol(data), ncol = ncol(data))
  colnames(p_matrix) <- var_names
  rownames(p_matrix) <- var_names
  
  for (i in seq_len(ncol(data))) {
    for (j in seq_len(ncol(data))) {
      if (i <= j) {
        test <- cor.test(data[[i]], data[[j]], method = "pearson")
        p_matrix[i, j] <- test$p.value
        p_matrix[j, i] <- test$p.value
      }
    }
  }
  
  list(cor_matrix = cor_matrix, p_matrix = p_matrix)
}




#### UI #### 

trend_choices <- c("Disease Structure Dynamics", "Thrombolysis Trend", "Mortality Trend", "Therapeutic Time Window")

data_choices <- c(
  "MICE_Ranger (with hospital)" = "data/data_summarized_by_month_filled_hosp_MR.tsv",
  "MICE (with hospital)" = "data/data_summarized_by_month_filled_hosp.tsv",
  "MICE" = "data/data_summarized_by_month_filled.tsv"
)


ui <- fluidPage(
  titlePanel("Vascular centers in St. Petersburg"),
  sidebarLayout(
    sidebarPanel(
      selectInput("trend", "Select indicator:", choices = trend_choices),
      selectInput("hospital", "Select hospital:", choices = c("All Hospitals", unique(data$hosp))),
      selectInput("type", "Select hospital type:", choices = c("All Types", unique(data$type))),
      selectInput("data_file", "Select data file:", choices = data_choices),
      checkboxInput("highlight_covid", "Highlight COVID period (Mar 2020 - Dec 2021)", value = FALSE),
      width = 3
    ),
    mainPanel(
      tabsetPanel(
        tabPanel("Yearly Dynamics", plotOutput("yearly_plot")),
        tabPanel("Monthly Dynamics", plotOutput("monthly_plot")),
        tabPanel("Correlation Matrix", plotOutput("correlation_matrix")),

      ),
      width = 9
    )
  )
)


#### Server #### 

server <- function(input, output) {

  # Reactive data processing
  data_reactive <- reactive({
    d <- load_data(input$data_file)
    d$type <- hosp_to_type[d$hosp]  
    d
  })

  # Yearly 
  
  output$yearly_plot <- renderPlot({
    data_filtered <- data_reactive()
    if (input$hospital != "All Hospitals") {
      data_filtered <- filter(data_filtered, hosp == input$hospital)
    }
    if (input$type != "All Types") {
      data_filtered <- filter(data_filtered, type == as.numeric(input$type))
    }

    
    if (input$trend == "Disease Structure Dynamics") {
      stroke_types <- c("ИИ_Ишемический_Инсульт", "ГИ_Геморрагический_Инсульт", "ТИА_Транзиторная_Ишемическая_Атака")
      plots <- lapply(stroke_types, function(stroke_type) {
        stroke_data <- data_filtered %>%
          group_by(year) %>%
          summarize(
            total_stroke = sum(!!sym(stroke_type), na.rm = TRUE),
            total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE)
          ) %>%
          mutate(Percentage = total_stroke / total_cases * 100)

        p <- ggplot(stroke_data, aes(x = year, y = Percentage)) +
          geom_line() +
          geom_point() +
          scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +
          scale_y_continuous(labels = scales::percent_format(scale = 1)) +
          labs(
            title = stroke_type,
            x = "",
            y = "% of total treated"
          ) +
          theme_custom

        # COVID period
        if (input$highlight_covid) {
          p <- p + annotate(
            "rect",
            xmin = 2020 + 2/12,  # Start of March 2020
            xmax = 2021 + 12/12, # End of December 2021
            ymin = -Inf,
            ymax = Inf,
            fill = "red",
            alpha = 0.1
          )
        }
        p
      })
      ggarrange(plotlist = plots, ncol = 3, nrow = 1)
      
    } else if (input$trend == "Thrombolysis Trend") {
      thromb_data <- data_filtered %>%
        group_by(year) %>%
        summarize(
          total_thromb = sum(Тромболизис, na.rm = TRUE),
          total_is = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE)
        ) %>%
        mutate(Percentage = total_thromb / total_is * 100)

      p <- ggplot(thromb_data, aes(x = year, y = Percentage)) +
        geom_line() +
        geom_point() +
        scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +
        scale_y_continuous(labels = scales::percent_format(scale = 1)) +
        labs(
          title = "Yearly Thrombolysis Trends",
          x = "",
          y = "% of IS Cases"
        ) +
        theme_custom

      
      if (input$highlight_covid) {
        p <- p + annotate(
          "rect",
          xmin = 2020 + 2/12,
          xmax = 2021 + 12/12,
          ymin = -Inf,
          ymax = Inf,
          fill = "red",
          alpha = 0.1
        )
      }
      p

    } else if (input$trend == "Mortality Trend") {
      mortality_data <- data_filtered %>%
        group_by(year) %>%
        summarize(
          total_deaths = sum(Летальные_Исходы, na.rm = TRUE),
          total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE)
        ) %>%
        mutate(Percentage = total_deaths / total_cases * 100)

      p <- ggplot(mortality_data, aes(x = year, y = Percentage)) +
        geom_line() +
        geom_point() +
        scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +
        scale_y_continuous(labels = scales::percent_format(scale = 1)) +
        labs(
          title = "Yearly Mortality Trends",
          x = "",
          y = "% of total treated"
        ) +
        theme_custom

      
      if (input$highlight_covid) {
        p <- p + annotate(
          "rect",
          xmin = 2020 + 2/12,
          xmax = 2021 + 12/12,
          ymin = -Inf,
          ymax = Inf,
          fill = "red",
          alpha = 0.1
        )
      }
      p
    } else if (input$trend == "Therapeutic Time Window") {
      therapeutic_window_data <- data_filtered %>%
        group_by(year) %>%
        summarize(
          total_therapeutic_window_perc = sum(Терапевтическое_Окно_4_5ч, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100
        ) %>%
        mutate(Percentage = total_therapeutic_window_perc)
      
        p <- ggplot(therapeutic_window_data, aes(x = year, y = Percentage)) +
        geom_line() +
        geom_point() +
        scale_x_continuous(breaks = seq(2017, 2024, by = 1)) +
        scale_y_continuous(labels = scales::percent_format(scale = 1)) +
        labs(
          title = "Yearly Therapeutic Time Window Trends",
          x = "",
          y = "% of total treated"
        ) +
        theme_custom

      
      if (input$highlight_covid) {
        p <- p + annotate(
          "rect",
          xmin = 2020 + 2/12,
          xmax = 2021 + 12/12,
          ymin = -Inf,
          ymax = Inf,
          fill = "red",
          alpha = 0.1
        )
      }
      p
}
  })


  # Monthly 
  
  
  output$monthly_plot <- renderPlot({
    data_filtered <- data_reactive()
    if (input$hospital != "All Hospitals") {
      data_filtered <- filter(data_filtered, hosp == input$hospital)
    }
    if (input$type != "All Types") {
      data_filtered <- filter(data_filtered, type == as.numeric(input$type))
    }

    if (input$trend == "Disease Structure Dynamics") {
      stroke_types <- c("ИИ_Ишемический_Инсульт", "ГИ_Геморрагический_Инсульт", "ТИА_Транзиторная_Ишемическая_Атака")
      plots <- lapply(stroke_types, function(stroke_type) {
        stroke_data <- data_filtered %>%
          mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
          group_by(year_month) %>%
          summarize(
            total_stroke = sum(!!sym(stroke_type), na.rm = TRUE),
            total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE)
          ) %>%
          mutate(Percentage = total_stroke / total_cases * 100)

        p <- ggplot(stroke_data, aes(x = year_month, y = Percentage)) +
          geom_line() +
          geom_point() +
          geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
          scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") +
          scale_y_continuous(labels = scales::percent_format(scale = 1)) +
          labs(
            title = stroke_type,
            x = "",
            y = "% of total treated"
          ) +
          theme_custom

        # Conditionally add highlight for COVID period
        if (input$highlight_covid) {
          p <- p + annotate(
            "rect",
            xmin = as.Date("2020-03-01"),
            xmax = as.Date("2021-12-31"),
            ymin = -Inf,
            ymax = Inf,
            fill = "red",
            alpha = 0.1
          )
        }
        p
      })
      ggarrange(plotlist = plots, ncol = 3, nrow = 1)
      
    } else if (input$trend == "Thrombolysis Trend") {
    thromb_data <- data_filtered %>%
      mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
      group_by(year_month) %>%
      summarize(
        total_thromb = sum(Тромболизис, na.rm = TRUE),
        total_is = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE)
      ) %>%
      mutate(Percentage = total_thromb / total_is * 100)

    p <- ggplot(thromb_data, aes(x = year_month, y = Percentage)) +
      geom_line() +
      geom_point() +
      geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
      scale_y_continuous(labels = scales::percent_format(scale = 1)) +
      scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") +
      labs(
        title = "Monthly Thrombolysis Trends",
        x = "",
        y = "% of IS Cases"
      ) +
      theme_custom

    
    if (input$highlight_covid) {
      p <- p + annotate(
        "rect",
        xmin = as.Date("2020-03-01"),
        xmax = as.Date("2021-12-31"),
        ymin = -Inf,
        ymax = Inf,
        fill = "red",
        alpha = 0.1
      )
    }
    p

  } else if (input$trend == "Mortality Trend") {
    mortality_data <- data_filtered %>%
      mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
      group_by(year_month) %>%
      summarize(
        total_deaths = sum(Летальные_Исходы, na.rm = TRUE),
        total_cases = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE)
      ) %>%
      mutate(Percentage = total_deaths / total_cases * 100)

    p <- ggplot(mortality_data, aes(x = year_month, y = Percentage)) +
      geom_line() +
      geom_point() +
      geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
      scale_y_continuous(labels = scales::percent_format(scale = 1)) +
      scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") +
      labs(
        title = "Monthly Mortality Trends",
        x = "",
        y = "% of total treated"
      ) +
      theme_custom

    
    if (input$highlight_covid) {
      p <- p + annotate(
        "rect",
        xmin = as.Date("2020-03-01"),
        xmax = as.Date("2021-12-31"),
        ymin = -Inf,
        ymax = Inf,
        fill = "red",
        alpha = 0.1
      )
    }
    p
  } else if (input$trend == "Therapeutic Time Window") {
    therapeutic_window_data <- data_filtered %>%
      mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
      group_by(year_month) %>%
      summarize(
          total_therapeutic_window_perc = sum(Терапевтическое_Окно_4_5ч, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100
        ) %>%
      mutate(Percentage = total_therapeutic_window_perc)

    p <- ggplot(therapeutic_window_data, aes(x = year_month, y = Percentage)) +
      geom_line() +
      geom_point() +
      geom_smooth(method = "loess", span = 0.5, se = TRUE, color = "darkred") +
      scale_y_continuous(labels = scales::percent_format(scale = 1)) +
      scale_x_date(date_breaks = "4 months", date_labels = "%Y-%m") +
      labs(
        title = "Monthly Therapeutic Time Window Trends",
        x = "Month",
        y = "% of total treated"
      ) +
      theme_custom

    
    if (input$highlight_covid) {
      p <- p + annotate(
        "rect",
        xmin = as.Date("2020-03-01"),
        xmax = as.Date("2021-12-31"),
        ymin = -Inf,
        ymax = Inf,
        fill = "red",
        alpha = 0.1
      )
    }
    p
  }
  })

  
  # Correlation matrix
  
  
  output$correlation_matrix <- renderPlot({
    data_filtered <- data_reactive()
    if (input$hospital != "All Hospitals") {
      data_filtered <- filter(data_filtered, hosp == input$hospital)
    }
    if (input$type != "All Types") {
      data_filtered <- filter(data_filtered, type == as.numeric(input$type))
    }

    data_prepared <- data_filtered %>%
      mutate(year_month = anytime::anydate(paste0(year, "-", sprintf("%02d", month), "-01"))) %>%
      group_by(year_month) %>%
    summarise(
      mortality = sum(Летальные_Исходы, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
      IS_perc = sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
      GS_perc = sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
      age60_perc = sum(Старше_60_лет, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
      severety_perc = sum(NIHSS_Больше_21, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
      ii_TW_perc = sum(ИИ_Терапевтическое_Окно, na.rm = TRUE) / sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) * 100,
      gi_TW_perc = sum(ГИ_Терапевтическое_Окно, na.rm = TRUE) / sum(ГИ_Геморрагический_Инсульт, na.rm = TRUE) * 100,
      total_window = sum(ИИ_Терапевтическое_Окно, ГИ_Терапевтическое_Окно, na.rm = TRUE) / sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE) * 100,
      thromb_perc = sum(Тромболизис, na.rm = TRUE) / sum(ИИ_Ишемический_Инсульт, na.rm = TRUE) * 100,
      total_treated = sum(Пролечены_с_ОНМК_Всего, na.rm = TRUE),
      total_load = sum(Суммарная_Загрузка_ОНМК, na.rm = TRUE),
      .groups = "drop"
    )

    results <- calc_cor_pval(data_prepared[,-1])
    cor_matrix <- results$cor_matrix
    p_matrix <- results$p_matrix

    cor_melted <- melt(cor_matrix)
    p_melted <- melt(p_matrix)

    cor_melted <- cor_melted %>%
      rename(correlation = value) %>%
      mutate(p_value = p_melted$value) %>%
      mutate(
        sig_label = case_when(
          p_value < 0.001 ~ "***",
          p_value < 0.01 ~ "**",
          p_value < 0.05 ~ "*",
          TRUE ~ ""
        )
      )

    ggplot(cor_melted, aes(Var1, Var2, fill = correlation)) +
      geom_tile(color = "white") +
      geom_text(aes(label = paste0(round(correlation, 2), sig_label)), color = "black", size = 5) +
      scale_fill_gradient2(low = "blue", high = "red", mid = "white", midpoint = 0, limit = c(-1, 1), space = "Lab", name = "Pearson r") +
      labs(title = "Correlation Matrix", x = "", y = "") +
      theme_minimal() +
      theme(axis.text.x = element_text(angle = 45, hjust = 1),
            axis.text = element_text(size = 12))
  })
}



#
shinyApp(ui = ui, server = server)



```
